# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Property Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  tenantId: number;
  tenantName: string;
  fileName: string;
  fileType: string;
  /** @nullable */
  filePath?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}