# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Property Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  name: string;
  address: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  postcode?: string | null;
  /** @nullable */
  notes?: string | null;
  totalRooms: number;
  occupiedRooms: number;
  expectedRent: number;
  createdAt: string;
}