# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Property Management API
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyUpdate {
  /** @minLength 1 */
  name?: string;
  address?: string;
  city?: string;
  postcode?: string;
  notes?: string;
}