# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Property Management API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomStatus } from './roomStatus';

export interface Room {
  id: number;
  propertyId: number;
  name: string;
  keyTag: string;
  status: RoomStatus;
  monthlyRent: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tenantId?: number | null;
  /** @nullable */
  tenantName?: string | null;
  createdAt: string;
}